Consolidate conftest.py
