+++ kevm-pyk/src/tests/conftest.py
@@ -3,9 +3,11 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import pytest
+from pyk.cli.utils import dir_path
 
 if TYPE_CHECKING:
-    from pytest import FixtureRequest, Parser
+    from pathlib import Path
+    from pytest import FixtureRequest, Parser, TempPathFactory
 
 
 def pytest_addoption(parser: Parser) -> None:
@@ -27,6 +29,11 @@ def pytest_addoption(parser: Parser) -> None:
         type=str,
         help='Run only this specific specification (skip others)',
     )
+    parser.addoption(
+        '--kompiled-targets-dir',
+        type=dir_path,
+        help='Use pre-kompiled definitions for proof tests',
+    )
 
 
 @pytest.fixture
@@ -42,3 +49,16 @@ def use_booster(request: FixtureRequest) -> bool:
 @pytest.fixture(scope='session')
 def spec_name(request: FixtureRequest) -> str | None:
     return request.config.getoption('--spec-name')
+
+
+@pytest.fixture(scope='session')
+def kompiled_targets_dir(request: FixtureRequest, tmp_path_factory: TempPathFactory) -> Path:
+    dir = request.config.getoption('--kompiled-targets-dir')
+    if dir:
+        return dir
+    else:
+        return tmp_path_factory.mktemp('prekompiled')
+
+
+def pytest_configure(config):
+    config.addinivalue_line("markers", "env(name): mark test to run only on named environment")
